submodule/bigquery/requirements.txt
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
coverage==7.3.2
db-dtypes==1.2.0
exceptiongroup==1.1.3
google-api-core==2.13.0
google-auth==2.23.3
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.13.0
google-cloud-core==2.3.3
google-cloud-secret-manager==2.16.4
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
idna==2.10
iniconfig==2.0.0
numpy==1.19.4
oauthlib==3.2.2
packaging==23.2
pandas==1.1.4
pandas-gbq==0.23.0
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==16.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydata-google-auth==1.8.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==2.0.0
rsa==4.6
six==1.15.0
tomli==2.0.1
urllib3==1.26.2
